removed circuit rpc functions and tests running fine
