+++ circuit/pallets/execution-delivery/rpc/Cargo.toml
@@ -17,6 +17,7 @@ codec = { package = "parity-scale-codec", version = "2.0.0" }
 jsonrpc-core = "15.0.0"
 jsonrpc-core-client = "15.0.0"
 jsonrpc-derive = "15.0.0"
+serde_json = "1.0.59"
 
 serde = { version = "1.0.124", features = ["derive"] }
